Refactor into Config
